+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [#97] - Gateway Infrastructure Setup (_by [@mathew-odwyer]_).
   - [#94] - Gateway Registration Flow (_by [@mathew-odwyer]_).
   - [#93] - Gateway Login Flow (_by [@mathew-odwyer]_).
+  - [#96] - Gateway Logout Flow (_by [@mathew-odwyer]_).
 
 ### Fixed
 
@@ -73,6 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v0.2.0]: https://github.com/mathew-odwyer/Server/releases/tag/v0.2.0
 [v0.1.0]: https://github.com/mathew-odwyer/Server/releases/tag/v0.0.1
 
+[#96]: https://github.com/mathew-odwyer/Server/issues/96
 [#93]: https://github.com/mathew-odwyer/Server/issues/93
 [#94]: https://github.com/mathew-odwyer/Server/issues/94
 [#97]: https://github.com/mathew-odwyer/Server/issues/97